// admin/src/modulos/api/model/farmacia.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Endereco } from './endereco';
import { FarmaciaCnae } from './farmaciaCnae';
import { FarmaciaEquipe } from './farmaciaEquipe';
import { ListaEmails } from './listaEmails';
import { ListaSites } from './listaSites';
import { ListaTelefones } from './listaTelefones';
import { RedesSociais } from './redesSociais';


export interface Farmacia {
    /**
     * código global único do objeto
     */
    id?: string;
    /**
     * Código CNPJ da farmácia
     */
    cnpj: string;
    /**
     * Nome do médico
     */
    nome: string;
    /**
     * Telefone principal dda farmácia
     */
    telefone?: string;
    /**
     * e-mail principal do médico, será usado para validações e informes
     */
    email: string;
    /**
     * Nome fantasia da empresa
     */
    nomeFantasia?: string;
    cnae?: FarmaciaCnae;
    /**
     * Número da Autorização de Funcionamento
     */
    numeroAutorizacao?: string;
    /**
     * Número da Licença de Funcionamento
     */
    licencaFuncionamento?: string;
    /**
     * Número da Inscrição Estadual
     */
    inscricaoEstadual?: string;
    endereco?: Endereco;
    telefones?: ListaTelefones;
    emails?: ListaEmails;
    sites?: ListaSites;
    redesSociais?: RedesSociais;
    /**
     * string base64 com o conteúdo de imagem da logo
     */
    logo?: string;
    equipe?: Array<FarmaciaEquipe>;
}