// admin/src/modulos/api/model/endereco.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Geolocalizacao } from './geolocalizacao';
import { ListaTelefones } from './listaTelefones';
import { UF } from './uF';


export interface Endereco {
    /**
     * Tipo do endereço, classifica o endereço de acordo com a finalidade
     */
    tipo?: Endereco.TipoEnum;
    /**
     * Parte longa do endereço como rua, avenida, quadra, praça
     */
    logradouro?: string;
    numero?: string;
    /**
     * Partes não descritas do endereço ou pontos de referencia
     */
    complemento?: string;
    /**
     * Bairro do endereço, favor enviar dado preferencialmente de acordo com o IBGE
     */
    bairro?: string;
    /**
     * cidade do endereço, favor enviar dado de acordo com a tabela do IBGE
     */
    cidade?: string;
    uf?: UF;
    /**
     * Código do CEP da localidade
     */
    cep?: string;
    telefones?: ListaTelefones;
    localizacao?: Geolocalizacao;
}
export namespace Endereco {
    export type TipoEnum = 'residencial' | 'consultorio' | 'atendimento';
    export const TipoEnum = {
        Residencial: 'residencial' as TipoEnum,
        Consultorio: 'consultorio' as TipoEnum,
        Atendimento: 'atendimento' as TipoEnum
    }
}