// admin/src/modulos/api/model/farmaciaEquipe.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DocumentoIdentidade } from './documentoIdentidade';


export interface FarmaciaEquipe {
    /**
     * Função do empregado na farmácia
     */
    funcao?: FarmaciaEquipe.FuncaoEnum;
    /**
     * Código CPF do Responsável pela Farmácia
     */
    cpf?: string;
    /**
     * Nome do responsável da farmácia
     */
    nome?: string;
    identidade?: DocumentoIdentidade;
    /**
     * Celular principal do responsável, será usado para validar ou assinar receitas
     */
    celular?: string;
    /**
     * e-mail principal do médico, será usado para validações e informes
     */
    email?: string;
    /**
     * ID único do usuário
     */
    user?: string;
}
export namespace FarmaciaEquipe {
    export type FuncaoEnum = 'responsavel' | 'atendente' | 'farmaceutico';
    export const FuncaoEnum = {
        Responsavel: 'responsavel' as FuncaoEnum,
        Atendente: 'atendente' as FuncaoEnum,
        Farmaceutico: 'farmaceutico' as FuncaoEnum
    }
}