// admin/src/modulos/api/model/cNAE.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CNAEClasse } from './cNAEClasse';
import { CNAEDivisao } from './cNAEDivisao';
import { CNAEGrupo } from './cNAEGrupo';
import { CNAESecao } from './cNAESecao';


export interface CNAE {
    secao?: CNAESecao;
    divisao?: CNAEDivisao;
    grupo?: CNAEGrupo;
    classe?: CNAEClasse;
    codigo?: string;
    codigoNumerico?: number;
    descricao?: string;
}